fix(middlewared): only copy skel if home dir exists
Ticket:	#25565

+++ src/middlewared/middlewared/plugins/account.py
@@ -194,14 +194,15 @@ class UserService(CRUDService):
 
         await self.__set_smbpasswd(data['username'], password)
 
-        for f in os.listdir(SKEL_PATH):
-            if f.startswith('dot'):
-                dest_file = os.path.join(data['home'], f[3:])
-            else:
-                dest_file = os.path.join(data['home'], f)
-            if not os.path.exists(dest_file):
-                shutil.copyfile(os.path.join(SKEL_PATH, f), dest_file)
-                os.chown(dest_file, data['uid'], group['gid'])
+        if os.path.exists(data['home']):
+            for f in os.listdir(SKEL_PATH):
+                if f.startswith('dot'):
+                    dest_file = os.path.join(data['home'], f[3:])
+                else:
+                    dest_file = os.path.join(data['home'], f)
+                if not os.path.exists(dest_file):
+                    shutil.copyfile(os.path.join(SKEL_PATH, f), dest_file)
+                    os.chown(dest_file, data['uid'], group['gid'])
 
         return pk
